🚚 Implement resistive TF coil geometry calculations in ResistiveTFCoil class and create new resistive tf coil test file
